refactor: move skeleton.Merge to kickoff package
